Fixing Student Report page
