--- views/partials/side-panel.ejs
@@ -1,18 +1,0 @@
-<nav id="sidebar">
-    <div class="p-4 pt-5">
-      <a href="#" class="img logo rounded-circle mb-5" style="background-image: url(images/logo.jpg);border:10px #008080"></a>
-<header id="stuModule-text"><B>FACULTY MODULE</B></header>
-<ul class="list-unstyled components mb-5">
-  <li class>
-    <a href="/faculty/awards">Awards</a>
-  </li>
-  <li>
-    <a href="/faculty/clubActivities">Club Activities</a>
-    </li>
-  
-  <li>
-  <a href="/faculty/eventsAttended">Events Attended</a>
-  </li>
-   </ul>
-</div>
-</nav>
