+++ views/partials/top-nav.ejs
@@ -2,16 +2,6 @@
     <div class="container-fluid">
       <div class="collapse navbar-collapse" id="navbarSupportedContent">
 
-        <!-- Toggle Nav Bar -->
-
-        <!-- <button type="button" id="sidebarCollapse" class="btn btn-primary">
-            <i class="fa fa-bars"></i>
-            <span class="sr-only">Toggle Menu</span>
-          </button>
-          <button class="btn btn-dark d-inline-block d-lg-none ml-auto" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-              <i class="fa fa-bars"></i>
-          </button> -->
-
         <div class="images">
           <img src="../img/college-logo.png" width="150" height="90">
         </div>
